work for split in process
