+++ js/main.js
@@ -30,165 +30,690 @@
   return r;
 })()({
   1: [function (require, module, exports) {
-    const mobileMenuIcon = document.querySelector('.mobile-menu');
-    const mobileMenu = document.querySelector('.mobile-navigation');
-    const closeMobMenu = document.querySelector('.close-mobile-nav');
+    require('./validator.js');
 
-    mobileMenuIcon.onclick = e => {
-      e.preventDefault();
-      mobileMenu.classList.add('show-nav');
-      setTimeout(() => {
-        mobileMenu.classList.add('visible-nav');
-      }, 10);
-    };
+    if (!!document.querySelector('.review__headline')) {
+      const mobileMenuIcon = document.querySelector('.mobile-menu');
+      const mobileMenu = document.querySelector('.mobile-navigation');
+      const closeMobMenu = document.querySelector('.close-mobile-nav');
 
-    closeMobMenu.onclick = e => {
-      e.preventDefault();
-      mobileMenu.classList.remove('visible-nav');
-      setTimeout(() => {
-        mobileMenu.classList.remove('show-nav');
-      }, 1000);
-    };
-  }, {}],
-  2: [function (require, module, exports) {
-    const name = document.getElementById('name');
-    const surName = document.getElementById('surname');
-    const tel = document.querySelector('#tel');
-    const errorMessageEmail = document.querySelector('.wrong-email-hidden');
-    const errorMessageTel = document.querySelector('.wrong-number-hidden');
-    const popUpContainer = document.querySelector('.pop-up-container');
-    const formInputs = document.getElementsByTagName('input');
-    const pushButton = document.querySelector('.push');
-
-    const VMasker = require('vanilla-masker');
-
-    const validate = require('validate-js');
-
-    const clearInputs = () => {
-      for (let i = 0; i < formInputs.length; i += 1) {
-        formInputs[i].value = '';
-      }
-    };
-
-    const showPopUpWindow = () => {
-      popUpContainer.classList.remove('pop-up-none');
-      popUpContainer.classList.add('pop-up-show');
-      const {
-        scrollX
-      } = window;
-      const {
-        scrollY
-      } = window;
-
-      window.onscroll = () => {
-        window.scrollTo(scrollX, scrollY);
+      mobileMenuIcon.onclick = e => {
+        e.preventDefault();
+        mobileMenu.classList.add('show-nav');
+        setTimeout(() => {
+          mobileMenu.classList.add('visible-nav');
+        }, 10);
       };
-    };
 
-    const clearErrors = () => {
-      name.classList.remove('review-user__input-error');
-      surName.classList.remove('review-user__input-error');
-      errorMessageTel.classList.remove('wrong-number-or-email-visible');
-      errorMessageEmail.classList.remove('wrong-number-or-email-visible');
-    };
-
-    const constraints = [{
-      name: 'name',
-      display: 'required',
-      rules: 'required|min_length[2]|alpha'
-    }, {
-      name: 'surname',
-      display: 'required',
-      rules: 'required|min_length[2]|alpha'
-    }, {
-      name: 'telephone',
-      display: 'Telephone No',
-      rules: 'required|callback_check_phone'
-    }, {
-      name: 'email',
-      display: 'Email No',
-      rules: 'required|valid_email'
-    }];
-    const validator = new validate('form', constraints, (errors, evt) => {
-      clearErrors();
-
-      if (errors.length > 0) {
-        for (let i = 0; i < errors.length; i++) {
-          if (errors[i].id === 'name') {
-            name.classList.add('review-user__input-error');
-          } else if (errors[i].id === 'surname') {
-            surName.classList.add('review-user__input-error');
-          } else if (errors[i].id === 'tel') {
-            errorMessageTel.classList.add('wrong-number-or-email-visible');
-          } else if (errors[i].id === 'email') {
-            errorMessageEmail.classList.add('wrong-number-or-email-visible');
-          }
-        }
-      } else {
-        clearInputs();
-        popUpContainer.classList.add('pop-up-show');
-        showPopUpWindow();
-      }
-
-      if (evt && evt.preventDefault) {
-        evt.preventDefault();
-      } else if (evt) {
-        evt.returnValue = false;
-      }
-    });
-
-    for (const input of formInputs) {
-      const currentInputName = input.name;
-      const currentInputSurName = input.surname;
-      const currentInputTel = input.telephone;
-      const currentInputEmail = input.email;
-      input.addEventListener('change', () => {
-        const click = document.createEvent('MouseEvents');
-        click.initMouseEvent('click', true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
-        constraints.forEach(item => {
-          if (item.name === currentInputName) {
-            pushButton.dispatchEvent(click);
-          } else if (item.name === currentInputSurName) {
-            pushButton.dispatchEvent(click);
-          } else if (item.name === currentInputTel) {
-            pushButton.dispatchEvent(click);
-          } else if (item.name === currentInputEmail) {
-            pushButton.dispatchEvent(click);
-          }
-        });
-      });
+      closeMobMenu.onclick = e => {
+        e.preventDefault();
+        mobileMenu.classList.remove('visible-nav');
+        setTimeout(() => {
+          mobileMenu.classList.remove('show-nav');
+        }, 1000);
+      };
+    } else {
+      return;
     }
-
-    validator.registerCallback('check_phone', value => {
-      const phoneCheck = /^(\s*)?(\+)?([- _():=+]?\d[- _():=+]?){10,14}(\s*)?$/;
-      return phoneCheck.test(value);
-    }).setMessage('check_phone', ''); // mask for phone
-
-    const inputHandler = (masks, max, event) => {
-      const input = event.target;
-      const inputValue = input.value.replace(/\D/g, '');
-      const maxLength = input.value.length > max ? 1 : 0;
-      VMasker(input).unMask();
-      VMasker(input).maskPattern(masks[maxLength]);
-      input.value = VMasker.toPattern(inputValue, masks[maxLength]);
-    };
-
-    const telMask = ['+9(999)999-99-99', '+9(999)999-99-99'];
-    VMasker(tel).maskPattern(telMask[0]);
-    tel.addEventListener('input', inputHandler.bind(undefined, telMask, 14), false); //
   }, {
-    "validate-js": 4,
-    "vanilla-masker": 5
+    "./validator.js": 2
+  }],
+  2: [function (require, module, exports) {
+    require('./animation.js');
+
+    if (!!document.querySelector('.review__headline')) {
+      const name = document.getElementById('name');
+      const surName = document.getElementById('surname');
+      const tel = document.querySelector('#tel');
+      const errorMessageEmail = document.querySelector('.wrong-email-hidden');
+      const errorMessageTel = document.querySelector('.wrong-number-hidden');
+      const popUpContainer = document.querySelector('.pop-up-container');
+      const formInputs = document.getElementsByTagName('input');
+      const pushButton = document.querySelector('.push');
+
+      const VMasker = require('vanilla-masker');
+
+      const validate = require('validate-js');
+
+      const clearInputs = () => {
+        for (let i = 0; i < formInputs.length; i += 1) {
+          formInputs[i].value = '';
+        }
+      };
+
+      const showPopUpWindow = () => {
+        popUpContainer.classList.remove('pop-up-none');
+        popUpContainer.classList.add('pop-up-show');
+        const {
+          scrollX
+        } = window;
+        const {
+          scrollY
+        } = window;
+
+        window.onscroll = () => {
+          window.scrollTo(scrollX, scrollY);
+        };
+      };
+
+      const clearErrors = () => {
+        name.classList.remove('review-user__input-error');
+        surName.classList.remove('review-user__input-error');
+        errorMessageTel.classList.remove('wrong-number-or-email-visible');
+        errorMessageEmail.classList.remove('wrong-number-or-email-visible');
+      };
+
+      const constraints = [{
+        name: 'name',
+        display: 'required',
+        rules: 'required|min_length[2]|alpha'
+      }, {
+        name: 'surname',
+        display: 'required',
+        rules: 'required|min_length[2]|alpha'
+      }, {
+        name: 'telephone',
+        display: 'Telephone No',
+        rules: 'required|callback_check_phone'
+      }, {
+        name: 'email',
+        display: 'Email No',
+        rules: 'required|valid_email'
+      }];
+      const validator = new validate('form', constraints, (errors, evt) => {
+        clearErrors();
+
+        if (errors.length > 0) {
+          for (let i = 0; i < errors.length; i++) {
+            if (errors[i].id === 'name') {
+              name.classList.add('review-user__input-error');
+            } else if (errors[i].id === 'surname') {
+              surName.classList.add('review-user__input-error');
+            } else if (errors[i].id === 'tel') {
+              errorMessageTel.classList.add('wrong-number-or-email-visible');
+            } else if (errors[i].id === 'email') {
+              errorMessageEmail.classList.add('wrong-number-or-email-visible');
+            }
+          }
+        } else {
+          clearInputs();
+          popUpContainer.classList.add('pop-up-show');
+          showPopUpWindow();
+        }
+
+        if (evt && evt.preventDefault) {
+          evt.preventDefault();
+        } else if (evt) {
+          evt.returnValue = false;
+        }
+      });
+
+      for (const input of formInputs) {
+        const currentInputName = input.name;
+        const currentInputSurName = input.surname;
+        const currentInputTel = input.telephone;
+        const currentInputEmail = input.email;
+        input.addEventListener('change', () => {
+          const click = document.createEvent('MouseEvents');
+          click.initMouseEvent('click', true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
+          constraints.forEach(item => {
+            if (item.name === currentInputName) {
+              pushButton.dispatchEvent(click);
+            } else if (item.name === currentInputSurName) {
+              pushButton.dispatchEvent(click);
+            } else if (item.name === currentInputTel) {
+              pushButton.dispatchEvent(click);
+            } else if (item.name === currentInputEmail) {
+              pushButton.dispatchEvent(click);
+            }
+          });
+        });
+      }
+
+      validator.registerCallback('check_phone', value => {
+        const phoneCheck = /^(\s*)?(\+)?([- _():=+]?\d[- _():=+]?){10,14}(\s*)?$/;
+        return phoneCheck.test(value);
+      }).setMessage('check_phone', ''); // mask for phone
+
+      const inputHandler = (masks, max, event) => {
+        const input = event.target;
+        const inputValue = input.value.replace(/\D/g, '');
+        const maxLength = input.value.length > max ? 1 : 0;
+        VMasker(input).unMask();
+        VMasker(input).maskPattern(masks[maxLength]);
+        input.value = VMasker.toPattern(inputValue, masks[maxLength]);
+      };
+
+      const telMask = ['+9(999)999-99-99', '+9(999)999-99-99'];
+      VMasker(tel).maskPattern(telMask[0]);
+      tel.addEventListener('input', inputHandler.bind(undefined, telMask, 14), false); //
+    } else {
+      return;
+    }
+  }, {
+    "./animation.js": 1,
+    "validate-js": 6,
+    "vanilla-masker": 7
   }],
   3: [function (require, module, exports) {
     require('./main-page/validator.js');
 
     require('./main-page/animation.js');
+
+    require('./slider/slider-images.js');
+
+    require('./slider/slider-pagination.js');
   }, {
     "./main-page/animation.js": 1,
-    "./main-page/validator.js": 2
+    "./main-page/validator.js": 2,
+    "./slider/slider-images.js": 4,
+    "./slider/slider-pagination.js": 5
   }],
   4: [function (require, module, exports) {
+    require('./slider-pagination.js');
+
+    if (!!document.querySelector('.slider-headline')) {
+      let slides = document.querySelectorAll('.slide-single');
+      const sliderBlock = document.querySelector('.slider');
+      let links = [];
+      const next = document.querySelector('#next-slide');
+      const previous = document.querySelector('#prev-slide'); // start function create images on pages
+
+      const showImage = () => {
+        findLinksImg();
+        imgIndex = 0;
+        const offset1 = 1;
+        const img = document.createElement('img');
+        const img1 = document.createElement('img');
+        const img2 = document.createElement('img');
+        img.src = links[links.length - 1];
+        img.classList.add('slide-single');
+        img.style.left = `${offset1 - 100}%`;
+        sliderBlock.appendChild(img);
+        img1.src = links[imgIndex];
+        img1.classList.add('slide-single');
+        img1.style.left = `${offset * 100}%`;
+        sliderBlock.appendChild(img1);
+        img2.src = links[imgIndex + 1];
+        img2.classList.add('slide-single');
+        img2.style.left = `${offset1 * 100}%`;
+        sliderBlock.appendChild(img2);
+      }; // find link active image after shift
+
+
+      const findIndexLastImg = arg => {
+        for (; imgIndex < links.length; imgIndex += 1) {
+          if (links[imgIndex] === arg.src) {
+            return imgIndex;
+          }
+        }
+      }; // add new image on the right after shift
+
+
+      const addNextImage = () => {
+        const rightImg = document.createElement('img');
+
+        if (imgIndex >= links.length - 1) {
+          imgIndex = 0;
+        } else {
+          imgIndex += 1;
+        }
+
+        rightImg.src = links[imgIndex];
+        rightImg.classList.add('slide-single');
+        rightImg.style.left = `${100}%`;
+        sliderBlock.appendChild(rightImg);
+        imgIndex = 0;
+      }; // add new image on the left after shift
+
+
+      const addPrevImage = () => {
+        const leftImg = document.createElement('img');
+
+        if (imgIndex <= 0) {
+          imgIndex = links.length - 1;
+        } else {
+          imgIndex -= 1;
+        }
+
+        leftImg.src = links[imgIndex];
+        leftImg.classList.add('slide-single');
+        leftImg.style.left = `${-100}%`;
+        sliderBlock.insertBefore(leftImg, sliderBlock.children[0]);
+        imgIndex = 0;
+      }; // image shift function to the left
+
+
+      const shiftImageLeft = () => {
+        offset = 0;
+        let step2 = 0;
+
+        for (let i = 0; i < slides.length; i += 1) {
+          if (step2 + 1 === slides.length) {
+            offset = 1;
+          } else {
+            offset = 0;
+            step2 += 1;
+          }
+
+          slides[i].style.left = `${offset * 100 - 100}%`;
+        }
+      }; // image shift function to the rights
+
+
+      const shiftImageRight = () => {
+        offset = 0;
+        let step3 = slides.length;
+
+        for (let i = slides.length - 1; i >= 0; i -= 1) {
+          if (step3 - 1 > 0) {
+            offset = 1;
+            step3 -= 1;
+          } else {
+            offset = 0;
+          }
+
+          slides[i].style.left = `${offset * 100}%`;
+        }
+      }; // active image index definition
+
+
+      const searchActiveImg = () => {
+        slides = document.querySelectorAll('.slide-single');
+
+        for (let i = 0; i < links.length; i += 1) {
+          if (links[i] === slides[1].src) {
+            indexActiveImg = i;
+          }
+        }
+      }; // adding "active" class to a button
+
+
+      const addClassButton = () => {
+        let indexActiveButton = 0;
+        searchActiveImg();
+        indexActiveImg += 1;
+        const buttonArr = document.querySelectorAll('.img-navigation');
+        const buttonMore = document.getElementById('moreImg');
+        const buttonLess = document.getElementById('lessImg');
+
+        for (; indexActiveButton < buttonArr.length; indexActiveButton += 1) {
+          buttonArr[indexActiveButton].classList.remove('active');
+
+          if (buttonMore) {
+            buttonMore.classList.remove('active');
+          }
+
+          if (buttonLess) {
+            buttonLess.classList.remove('active');
+          }
+
+          if (indexActiveImg > +buttonArr[buttonArr.length - 1].innerText) {
+            buttonMore.classList.add('active');
+          }
+
+          if (indexActiveImg < +buttonArr[0].innerText) {
+            buttonLess.classList.add('active');
+          } else if (+buttonArr[indexActiveButton].innerText === indexActiveImg) {
+            buttonArr[indexActiveButton].classList.add('active');
+          }
+        }
+
+        indexActiveImg = 0;
+      }; // function showing new image after pressing a button "next"
+
+
+      next.onclick = () => {
+        if (!isFlagAction) {
+          slides = document.querySelectorAll('.slide-single');
+          findIndexLastImg(slides[2]);
+          slides[0].remove();
+          shiftImageLeft();
+          isFlagAction = true;
+          setTimeout(() => {
+            addNextImage();
+            addClassButton();
+            isFlagAction = false;
+          }, 2000);
+        }
+
+        return false;
+      }; // function showing new image after pressing a button "previous"
+
+
+      previous.onclick = () => {
+        if (!isFlagAction) {
+          slides = document.querySelectorAll('.slide-single');
+          offset = 0;
+          findIndexLastImg(slides[0]);
+          slides[2].remove();
+          shiftImageRight();
+          isFlagAction = true;
+          setTimeout(() => {
+            addPrevImage();
+            addClassButton();
+            isFlagAction = false;
+          }, 2000);
+        }
+
+        return false;
+      };
+
+      showImage();
+    } else {
+      return;
+    }
+  }, {
+    "./slider-pagination.js": 5
+  }],
+  5: [function (require, module, exports) {
+    require('./slider-images.js');
+
+    if (!!document.querySelector('.slider-headline')) {
+      const paginationImgBlock = document.querySelector('.pagination-img-block');
+      let links = [];
+      let imgIndex = 0;
+      let offset = 0;
+      let indexPagination = 3; //  number of buttons pagination
+
+      let indexStartNumButton = 0;
+      let indexActiveImg = 0;
+      let numberPushButton = 0;
+      let activeButton = 0;
+      let linkNewImage = 0;
+      let isFlagAction = false;
+
+      const showButtonMore = () => {
+        const moreImg = document.createElement('button');
+        moreImg.classList.add('moreImg-class');
+        moreImg.id = 'moreImg';
+        moreImg.textContent = '>>';
+        paginationImgBlock.appendChild(moreImg);
+      };
+
+      const showButtonLess = () => {
+        const lessImg = document.createElement('button');
+        lessImg.classList.add('lessImg-class');
+        lessImg.id = 'lessImg';
+        lessImg.textContent = '<<';
+        paginationImgBlock.appendChild(lessImg);
+      }; // creating a new button with a number greater than the previous one
+
+
+      const createNewButtonMore = () => {
+        const imgNumber = document.createElement('button');
+        imgNumber.classList.add('img-navigation');
+        imgNumber.textContent = +indexStartNumButton + 1;
+        paginationImgBlock.appendChild(imgNumber);
+      }; // creating a new button with a number less than the previous one
+
+
+      const createNewButtonLess = () => {
+        const imgNumber = document.createElement('button');
+        imgNumber.classList.add('img-navigation');
+        imgNumber.textContent = +indexStartNumButton - 1;
+        paginationImgBlock.insertBefore(imgNumber, paginationImgBlock.children[0]);
+      }; // start function create pagination button
+
+
+      const showPagination = () => {
+        if (links.length > indexPagination) {
+          for (let i = 0; i < indexPagination; i += 1) {
+            const imgNumber = document.createElement('button');
+            imgNumber.classList.add('img-navigation');
+            imgNumber.textContent = i + 1;
+            paginationImgBlock.appendChild(imgNumber);
+          }
+
+          showButtonMore();
+        } else if (links.length <= indexPagination) {
+          for (let i = 0; i < links.length; i += 1) {
+            const imgNumber = document.createElement('button');
+            imgNumber.classList.add('img-navigation');
+            imgNumber.textContent = i + 1;
+            paginationImgBlock.appendChild(imgNumber);
+          }
+        }
+
+        addClassButton();
+      }; // search for the number of the last button pagination
+
+
+      const searchLastNumberButton = n => {
+        const buttonArr = document.querySelectorAll('.img-navigation');
+
+        for (let i = 0; i < buttonArr.length; i += 1) {
+          if (i + 1 === n) {
+            indexStartNumButton = +buttonArr[i].innerText;
+          }
+        }
+      }; // clear all button pagination
+
+
+      const clearPagination = () => {
+        while (paginationImgBlock.firstChild) {
+          paginationImgBlock.removeChild(paginationImgBlock.firstChild);
+        }
+      }; // show new pagination button after push button ">>"
+
+
+      const showNextPagination = () => {
+        searchLastNumberButton(indexPagination);
+        clearPagination();
+        showButtonLess();
+        const reserveVariables = indexStartNumButton;
+        indexPagination = indexStartNumButton + indexPagination;
+
+        if (indexPagination === links.length) {
+          for (indexStartNumButton; indexStartNumButton < indexPagination; indexStartNumButton += 1) {
+            createNewButtonMore();
+          }
+        } else if (indexPagination < links.length) {
+          for (indexStartNumButton; indexStartNumButton < indexPagination; indexStartNumButton += 1) {
+            createNewButtonMore();
+          }
+
+          showButtonMore();
+        } else {
+          for (indexStartNumButton; indexStartNumButton < links.length; indexStartNumButton += 1) {
+            createNewButtonMore();
+          }
+        }
+
+        indexPagination -= reserveVariables;
+        addClassButton();
+      }; // show new pagination button after push button "<<"
+
+
+      const showPrevPagination = () => {
+        searchLastNumberButton(1);
+        clearPagination();
+        showButtonMore();
+        const reserveVariables = indexStartNumButton;
+        indexPagination = indexStartNumButton - indexPagination;
+
+        if (indexPagination > 1) {
+          for (indexStartNumButton; indexStartNumButton > indexPagination; indexStartNumButton -= 1) {
+            createNewButtonLess();
+          }
+
+          const lessImg = document.createElement('button');
+          lessImg.classList.add('lessImg-class');
+          lessImg.id = 'lessImg';
+          lessImg.textContent = '<<';
+          paginationImgBlock.insertBefore(lessImg, paginationImgBlock.children[0]);
+        } else {
+          for (indexStartNumButton; indexStartNumButton > 1; indexStartNumButton -= 1) {
+            createNewButtonLess();
+          }
+        }
+
+        indexPagination = reserveVariables - indexPagination;
+        addClassButton();
+      }; // brute force pagination array
+
+
+      const bruteForcePagination = () => {
+        const buttonArr = document.querySelectorAll('.img-navigation');
+        const lessButton = document.getElementById('lessImg');
+        const moreButton = document.getElementById('moreImg');
+
+        if (lessButton) {
+          if (lessButton.classList.contains('active')) {
+            activeButton = -1;
+          }
+        }
+
+        if (moreButton) {
+          if (moreButton.classList.contains('active')) {
+            activeButton = links.length + 1;
+          }
+        }
+
+        for (let i = 0; i < buttonArr.length; i += 1) {
+          if (buttonArr[i].classList.contains('active')) {
+            activeButton = +buttonArr[i].innerText;
+          }
+        }
+      }; // search for a link to a picture of the corresponding pressed button
+
+
+      const searchIndexNewImage = () => {
+        for (let i = 0; i < links.length; i += 1) {
+          if (i + 1 === numberPushButton) {
+            linkNewImage = links[i];
+            break;
+          }
+        }
+      }; // adding a new image after pressing the button with the number more than the active one
+
+
+      const addNextImageFromPagination = () => {
+        if (isFlagAction) {
+          return;
+        }
+
+        const rightImg = document.createElement('img');
+        searchIndexNewImage();
+        slides[0].remove();
+        slides[2].remove();
+        rightImg.src = linkNewImage;
+        rightImg.classList.add('slide-single');
+        rightImg.style.left = `${100}%`;
+        sliderBlock.appendChild(rightImg);
+        slides = document.querySelectorAll('.slide-single');
+        isFlagAction = true;
+        setTimeout(() => {
+          shiftImageLeft();
+        }, 200);
+        setTimeout(() => {
+          findIndexLastImg(slides[1]);
+          addNextImage();
+          slides[0].remove();
+          findIndexLastImg(slides[1]);
+          addPrevImage();
+          addClassButton();
+          numberPushButton = 0;
+          linkNewImage = 0;
+          isFlagAction = false;
+        }, 2000);
+      }; // adding a new image after pressing the button with the number less than the active one
+
+
+      const addPrevImageFromPagination = () => {
+        if (isFlagAction) {
+          return;
+        }
+
+        const leftImg = document.createElement('img');
+        searchIndexNewImage();
+        slides[0].remove();
+        slides[2].remove();
+        leftImg.src = linkNewImage;
+        leftImg.classList.add('slide-single');
+        leftImg.style.left = `${-100}%`;
+        sliderBlock.insertBefore(leftImg, sliderBlock.children[0]);
+        slides = document.querySelectorAll('.slide-single');
+        isFlagAction = true;
+        setTimeout(() => {
+          shiftImageRight();
+        }, 200);
+        setTimeout(() => {
+          findIndexLastImg(slides[0]);
+          addPrevImage();
+          slides[1].remove();
+          findIndexLastImg(slides[0]);
+          addNextImage();
+          addClassButton();
+          numberPushButton = 0;
+          linkNewImage = 0;
+          isFlagAction = false;
+        }, 2000);
+      }; // showing a new image after pressing a button pagination
+
+
+      const showNotActiveImg = () => {
+        let indexWaitImage = 0;
+        const buttonArr = document.querySelectorAll('.img-navigation');
+        bruteForcePagination();
+
+        for (; indexWaitImage <= links.length; indexWaitImage += 1) {
+          if (indexWaitImage === numberPushButton) {
+            for (let i = 0; i < buttonArr.length; i += 1) {
+              if (+buttonArr[i].innerText === numberPushButton) {
+                if (indexWaitImage > activeButton) {
+                  addNextImageFromPagination();
+                  break;
+                } else {
+                  addPrevImageFromPagination();
+                  break;
+                }
+              }
+            }
+          }
+        }
+      }; // checking whether the active class has a button
+
+
+      const containsActiveClass = x => {
+        const numButton = +x;
+        const buttonArr = document.querySelectorAll('.img-navigation');
+
+        for (let i = 0; i < buttonArr.length; i += 1) {
+          if (numButton === +buttonArr[i].innerText) {
+            if (buttonArr[i].classList.contains('active')) {
+              break;
+            } else {
+              showNotActiveImg();
+              break;
+            }
+          }
+        }
+      };
+
+      paginationImgBlock.onclick = event => {
+        const {
+          target
+        } = event;
+
+        if (target.id === 'moreImg') {
+          showNextPagination();
+        } else if (target.id === 'lessImg') {
+          showPrevPagination();
+        } else if (target.id !== 'lessImg' && 'moreImg') {
+          numberPushButton = +target.innerText;
+          containsActiveClass(target.innerText);
+        }
+      };
+
+      showPagination();
+    } else {
+      return;
+    }
+  }, {
+    "./slider-images.js": 4
+  }],
+  6: [function (require, module, exports) {
     /*
      * validate.js 2.0.1
      * Copyright (c) 2011 - 2015 Rick Harrison, http://rickharrison.me
@@ -820,7 +1345,7 @@
       module.exports = FormValidator;
     }
   }, {}],
-  5: [function (require, module, exports) {
+  7: [function (require, module, exports) {
     (function (root, factory) {
       if (typeof define === 'function' && define.amd) {
         define(factory);
