+++ js/script.js
@@ -1,2 +1,4 @@
 require('./main-page/validator.js');
 require('./main-page/animation.js');
+require('./slider/slider-images.js');
+require('./slider/slider-pagination.js');
