+++ slider.html
@@ -92,7 +92,7 @@
         <span class="producer__text">Разработано</span>
     </div>
 </footer>
-<script src="js/slider/slider.js"></script>
+<script src="js/main.js"></script>
 </body>
 
 </html>
